// src/uiLibrary/enums/EnumPaletteShade.ts
enum EnumPaletteShade {
  shade50 = "50",
  shade100 = "100",
  shade150 = "150",
  shade200 = "200",
  shade250 = "250",
  shade300 = "300",
  shade350 = "350",
  shade400 = "400",
  shade450 = "450",
  shade500 = "500",
  shade550 = "550",
  shade600 = "600",
  shade650 = "650",
  shade700 = "700",
  shade750 = "750",
  shade800 = "800",
  shade850 = "850",
  shade900 = "900",
  shade950 = "950",
}

export default EnumPaletteShade;