// src/uiLibrary/interfaces/properties/IPropPaletteColour.ts
export default interface IPropPaletteColour {
  slate?: boolean;
  gray?: boolean;
  zinc?: boolean;
  neutral?: boolean;
  stone?: boolean;
  red?: boolean;
  orange?: boolean;
  amber?: boolean;
  yellow?: boolean;
  lime?: boolean;
  green?: boolean;
  emerald?: boolean;
  teal?: boolean;
  cyan?: boolean;
  sky?: boolean;
  blue?: boolean;
  indigo?: boolean;
  violet?: boolean;
  purple?: boolean;
  fuchsia?: boolean;
  pink?: boolean;
  rose?: boolean;
}