// src/uiLibrary/interfaces/properties/IPropPaletteShade.ts
export default interface IPropPaletteShade {
  shade50?: boolean;
  shade100?: boolean;
  shade150?: boolean;
  shade200?: boolean;
  shade250?: boolean;
  shade300?: boolean;
  shade350?: boolean;
  shade400?: boolean;
  shade450?: boolean;
  shade500?: boolean;
  shade550?: boolean;
  shade600?: boolean;
  shade650?: boolean;
  shade700?: boolean;
  shade750?: boolean;
  shade800?: boolean;
  shade850?: boolean;
  shade900?: boolean;
  shade950?: boolean;
}